Always use an NSString as helper for OC_PythonUnicode
* When sizeof(Py_UNICODE) == sizeof(unichar) we could work without the NSString,
  but for some reason that causes problems with some CFString APIs

* Provide a number of __realObject__ implementations to optimize the
  creation of the NSString: don't copy actual string data unless
  really necessary

* This checkin also adds caching for selector lookups, should slightly
  improve performance

This also add some other changes to reduce the amount of warnings with
newer versions of clang.

+++ pyobjc-core/Modules/objc/test/exceptions.m
@@ -8,7 +8,7 @@
 
 static NSString* addSomeUnicode(NSString* input)
 {
-	return [NSString stringWithFormat:@"%@%C%C", input, 0x1234, 0x2049];
+	return [NSString stringWithFormat:@"%@%C%C", input, (short)0x1234, (short)0x2049];
 }
 
 @interface PyObjCTestExceptions : NSObject
